refactor(reporting): move admin_registry to admin_registry.py
Split the source- and metric-registry admin routes
(reporting_sources_admin, api_admin_sources_list/create/update/delete,
reporting_metrics_admin, api_admin_metrics_list/create/update/delete),
their payload validators (_validate_source_payload,
_validate_metric_payload, _columns_to_json, _source_insert_params,
_metric_insert_params), and the two registry-cache invalidators
(invalidate_reporting_sources, invalidate_reporting_metrics) out of
nx_lib/views/reporting/__init__.py into admin_registry.py, part of
beautify-phase-2a Task 3.

invalidate_reporting_sources/metrics stay re-exported (and in
__all__) from __init__.py -- ops/run_scheduled_reports.py and
tests/unit/test_reporting_registry_cache.py both import them from
nx_lib.views.reporting directly. __init__.py delegates route
registration to admin_registry.register_routes(app). No test patches
any of the moved names by string, so no patch retargeting was needed
-- verified with the full test_reporting_routes.py and
test_reporting_registry_cache.py runs.
